% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency.R
\name{aextras_dependency}
\alias{aextras_dependency}
\alias{aextras_dep}
\title{Dependency}
\usage{
aextras_dependency(cdn = FALSE)

aextras_dep(cdn = FALSE)
}
\arguments{
\item{cdn}{Whether to use local or CDN files.}
}
\description{
Add extras dependency.
}
\examples{
library(aframer)

browse_aframe(
  a_scene(
    a_dependency(cdn = TRUE),
    aextras_dependency(cdn = TRUE)
  )
)

}
