% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{a_ocean}
\alias{a_ocean}
\alias{a_grid}
\alias{a_tube}
\alias{a_hexmap}
\alias{a_animation}
\title{Primitives}
\usage{
a_ocean(...)

a_grid(...)

a_tube(...)

a_hexmap(...)

a_animation(...)
}
\arguments{
\item{...}{Any aframer element or named options.}
}
\description{
Extra primitives.
}
\note{
Require dependency, see \code{\link{aextras_dependency}}.
}
\examples{
library(aframer)

browse_aframe(
  a_scene(
    a_dependency(),
    aextras_dependency(),
    a_ocean(
      width="50", depth="50", density="40"
    )
  )
)

}
