% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{a_ocean}
\alias{a_ocean}
\title{Primitives}
\usage{
a_ocean(..., cdn = FALSE, version = "0.8.0")
}
\arguments{
\item{...}{Any aframer element or named options.}

\item{cdn}{Whether to use local or CDN files.}

\item{version}{Passed to \link[aframer]{a_dependency}.}
}
\description{
Extra primitives.
}
\examples{
library(aframer)

browse_aframe(
  a_scene(
    a_ocean(
      width="50", depth="50", density="40"
    )
  )
)

}
